Improvements to javascript and build process
Added prettier for consistent code formatting and applied formatting.
(This accounts for most of the code changes)

Updated javascript to simplify some of the environment detection

Moved import of pako to top of file

Ensure that performance object in node is not fully replaced
and instead just `now` is replaced

Added scripts to package.json for building and testing.

Added source maps - this is nice for library consumers that need to debug

Updated compatibility chart. globalThis bumped node from 8 to 12.

Updated build process to automatically include type definition